refactor(cli): 拆分 ts_bridge 归档转换与 host 回调分发
- 归档 JSON 与 tool UI 映射提取为独立模块
- host JSON-RPC 回调迁入 host_dispatch.rs
